# swagger.yaml
swagger: '2.0'
info:
  description: "Milla is mostly Qt Jira-specific orchestrator service for the OpenReq Qt trial that operates between Qt Jira and the OpenReq infrastructure. Milla has five main functionalities divided into three different controllers. <br>\r\n<b>Milla controller:</b><br>1) Milla fetches all or updated data from Qt Jira and stores in a caching manner the data in the Mallikas database service. <br>2) Milla orchestrates the communication to Mallikas database service so that a project, requirement etc. can be manipulated (CRUD manner) in Mallikas database service in OpenReq JSON format.<br>3) Milla orchestrates the data from Mallikas database service to Mulperi service in order to manage dependencies, carry out analyses etc. \n<br><b>Detection controller:</b><br>4) Milla calls the natural language processing services of OpenReq infrastructre that can detect from the requirements text hidden or implicit dependencies to other requirements.  <br>\r\n<b>Qt controller:</b><br>5) Milla provides the access point for the Qt Jira dependency browser in the Qt trial. <br>\r\n"
  version: '1.13'
  title: Milla
host: 'localhost:9203'
basePath: /
...
paths:
  /detectSimilarityAddReqs:
    post:
      tags:
        - detection-controller
      summary: Cache requirements for UPC dependency Detection
      description: '<b>Functionality</b>: Post all requirements and dependencies in a project as a String to UPC services in order to be cached for dependency detection purposes. <br><b>Precondition</b>: The project has been cached in Mallikas.<br><b>Postcondition</b>: After successfully caching requirements in UPC service, similarity detection can be carried out.<br><br><b>Parameter: </b><br>projectId: The project id in Mallikas (e.g., QTWB).'
      operationId: postRequirementsToUPCSimilarityDetectionUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: projectId
          description: projectId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /detectSimilarityProject:
    post:
      tags:
        - detection-controller
      summary: Detect similarity between all requirements of a project using UPC Similarity Detection
      description: '<b>Functionality</b>: All requirements of a given project are posted to UPC Similarity Detection in order to detect similarity between all requirements. <br><b>Precondition</b>: The project has been cached in Mallikas.<br><b>Postcondition</b>: After successful detection, the detected new similarities are stored in Mallikas using the similarity dependency type and proposed status.<br><b>Parameters:</b><br>compare: Whether text attribute is used in comparison<br>projectId: The project id in Mallikas.<br>threshold: The minimum score for similarity detection (e.g. 0.3).'
      operationId: postRequirementsToUPCSimilarityDetectionProjectUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: compare
          in: query
          description: compare
          required: true
          type: boolean
        - name: projectId
          in: query
          description: projectId
          required: true
          type: string
        - name: threshold
          in: query
          description: threshold
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /detectSimilarityReqProject:
    post:
      tags:
        - detection-controller
      summary: Detect similarity of one requirement against all other requirements of a project using UPC Similarity Detection
      description: '<b>Functionality</b>: All requirements of a given project are posted to UPC Similarity Detection in order to detect similarity between one specified requirements in to project to all other requirements. <br><b>Precondition</b>: The project has been cached in Mallikas.<br><b>Postcondition</b>: After successful detection, the detected new similarities are stored in Mallikas using the similarity dependency type and proposed status.<br><b>Parameters:</b><br>compare: Whether the text attribute is used in comparison<br>projectId: The project id in Mallikas.<br>reqIds: The ids of the requirements that are to be compared to other requirements in the project.<br>threshold: The minimum score for similarity detection (e.g. 0.3).'
      operationId: postRequirementsToUPCSimilarityDetectionReqProjectUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: compare
          in: query
          description: compare
          required: true
          type: boolean
        - name: projectId
          in: query
          description: projectId
          required: true
          type: string
        - name: requirementId
          in: query
          description: requirementId
          required: true
          type: array
          items:
            type: string
          collectionFormat: multi
        - name: threshold
          in: query
          description: threshold
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /detectSimilarityReqReq:
    post:
      tags:
        - detection-controller
      summary: Detect similarity between two requirements using UPC Similarity Detection
      description: '<b>Functionality</b>: Two given requirements are posted to UPC Similarity Detection in order to detect similarity between them.<br><b>Precondition</b>: The two requirements have been cached in Mallikas.<br><b>Postcondition</b>: After successful detection, the similarity, if above treshold, is stored in Mallikas using the similarity dependency type and proposed status.<br><b>Parameters:</b><br>reqId1: The id of the requirement that is compared to other requirement (reqId2).<br>reqId2: The id of the requirement that is compared to other requirement (reqId1).'
      operationId: postRequirementsToUPCSimilarityDetectionReqReqUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: compare
          in: query
          description: compare
          required: true
          type: boolean
        - name: requirementId1
          in: query
          description: requirementId1
          required: true
          type: string
        - name: requirementId2
          in: query
          description: requirementId2
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
  /otherDetectionService:
    post:
      tags:
        - detection-controller
      summary: Post requirements to some detection service
      description: '<b>Functionality</b>: Post all requirements and dependencies to some detection service. <br><b>Precondition</b>: The project has been cached in Mallikas.<br><b>Postcondition</b>: Successfully detected dependencies are saved in Mallikas.<br><br>projectId: The project id in Mallikas (e.g., QTWB).<br>url: The url of the service to be used.'
      operationId: postRequirementsToDetectionServiceUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: projectId
          in: query
          description: projectId
          required: true
          type: string
        - name: url
          in: query
          description: url
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...